+++ fibonacci/odin.odin
@@ -2,7 +2,7 @@ package main
 
 import "core:fmt"
 
-fibonacci :: proc(n: int32) -> int32 {
+fibonacci :: proc(n: i64) -> i64 {
     if n <= 1 {
         return n;
     }
