+++ fibonacci/v.v
@@ -2,7 +2,7 @@ fn main() {
 	println(fib(47))
 }
 
-fn fib(n int) int {
+fn fib(n i64) i64 {
 	if n <= 1 { return n }
 	return fib(n - 1) + fib(n - 2)
 }
